+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- TUI: `c` key to copy selected job's command to clipboard
+
 ## [0.10.0] - 2025-12-01
 
 ### Added